+++ bank-feeds/src/sdk/sdk.ts
@@ -58,9 +58,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "3.0.0";
-    genVersion = "2.195.2";
-    userAgent = "speakeasy-sdk/typescript 3.0.0 2.195.2 3.0.0 @codat/bank-feeds";
+    sdkVersion = "3.0.1";
+    genVersion = "2.209.0";
+    userAgent = "speakeasy-sdk/typescript 3.0.1 2.209.0 3.0.0 @codat/bank-feeds";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);